Apisimilar/requirements.txt
fastapi
uvicorn
fasttext-wheel
pydantic
